Split some source files
